+++ raft/log.go
@@ -3,7 +3,7 @@ package raft
 import "fmt"
 
 const (
-	Normal int = iota
+	Normal int64 = iota
 
 	AddNode
 	RemoveNode
@@ -14,8 +14,8 @@ const (
 )
 
 type Entry struct {
-	Type int
-	Term int
+	Type int64
+	Term int64
 	Data []byte
 }
 
@@ -25,13 +25,13 @@ func (e *Entry) isConfig() bool {
 
 type log struct {
 	ents      []Entry
-	committed int
-	applied   int
-	offset    int
+	committed int64
+	applied   int64
+	offset    int64
 
 	// want a compact after the number of entries exceeds the threshold
 	// TODO(xiangli) size might be a better criteria
-	compactThreshold int
+	compactThreshold int64
 }
 
 func newLog() *log {
@@ -43,7 +43,7 @@ func newLog() *log {
 	}
 }
 
-func (l *log) maybeAppend(index, logTerm, committed int, ents ...Entry) bool {
+func (l *log) maybeAppend(index, logTerm, committed int64, ents ...Entry) bool {
 	if l.matchTerm(index, logTerm) {
 		l.append(index, ents...)
 		l.committed = committed
@@ -52,23 +52,23 @@ func (l *log) maybeAppend(index, logTerm, committed int, ents ...Entry) bool {
 	return false
 }
 
-func (l *log) append(after int, ents ...Entry) int {
+func (l *log) append(after int64, ents ...Entry) int64 {
 	l.ents = append(l.slice(l.offset, after+1), ents...)
 	return l.lastIndex()
 }
 
-func (l *log) lastIndex() int {
-	return len(l.ents) - 1 + l.offset
+func (l *log) lastIndex() int64 {
+	return int64(len(l.ents)) - 1 + l.offset
 }
 
-func (l *log) term(i int) int {
+func (l *log) term(i int64) int64 {
 	if e := l.at(i); e != nil {
 		return e.Term
 	}
 	return -1
 }
 
-func (l *log) entries(i int) []Entry {
+func (l *log) entries(i int64) []Entry {
 	// never send out the first entry
 	// first entry is only used for matching
 	// prevLogTerm
@@ -78,19 +78,19 @@ func (l *log) entries(i int) []Entry {
 	return l.slice(i, l.lastIndex()+1)
 }
 
-func (l *log) isUpToDate(i, term int) bool {
+func (l *log) isUpToDate(i, term int64) bool {
 	e := l.at(l.lastIndex())
 	return term > e.Term || (term == e.Term && i >= l.lastIndex())
 }
 
-func (l *log) matchTerm(i, term int) bool {
+func (l *log) matchTerm(i, term int64) bool {
 	if e := l.at(i); e != nil {
 		return e.Term == term
 	}
 	return false
 }
 
-func (l *log) maybeCommit(maxIndex, term int) bool {
+func (l *log) maybeCommit(maxIndex, term int64) bool {
 	if maxIndex > l.committed && l.term(maxIndex) == term {
 		l.committed = maxIndex
 		return true
@@ -112,27 +112,27 @@ func (l *log) nextEnts() (ents []Entry) {
 // i must be not smaller than the index of the first entry
 // and not greater than the index of the last entry.
 // the number of entries after compaction will be returned.
-func (l *log) compact(i int) int {
+func (l *log) compact(i int64) int64 {
 	if l.isOutOfBounds(i) {
 		panic(fmt.Sprintf("compact %d out of bounds [%d:%d]", i, l.offset, l.lastIndex()))
 	}
 	l.ents = l.slice(i, l.lastIndex()+1)
 	l.offset = i
-	return len(l.ents)
+	return int64(len(l.ents))
 }
 
 func (l *log) shouldCompact() bool {
 	return (l.applied - l.offset) > l.compactThreshold
 }
 
-func (l *log) restore(index, term int) {
+func (l *log) restore(index, term int64) {
 	l.ents = []Entry{{Term: term}}
 	l.committed = index
 	l.applied = index
 	l.offset = index
 }
 
-func (l *log) at(i int) *Entry {
+func (l *log) at(i int64) *Entry {
 	if l.isOutOfBounds(i) {
 		return nil
 	}
@@ -140,7 +140,7 @@ func (l *log) at(i int) *Entry {
 }
 
 // slice get a slice of log entries from lo through hi-1, inclusive.
-func (l *log) slice(lo int, hi int) []Entry {
+func (l *log) slice(lo int64, hi int64) []Entry {
 	if lo >= hi {
 		return nil
 	}
@@ -150,7 +150,7 @@ func (l *log) slice(lo int, hi int) []Entry {
 	return l.ents[lo-l.offset : hi-l.offset]
 }
 
-func (l *log) isOutOfBounds(i int) bool {
+func (l *log) isOutOfBounds(i int64) bool {
 	if i < l.offset || i > l.lastIndex() {
 		return true
 	}
